+++ examples/typescript/simple_sponsored_transaction.ts
@@ -70,17 +70,14 @@ const example = async () => {
     },
   });
 
-  // Alice signs
-  const senderSignature = aptos.transaction.sign({ signer: alice, transaction });
-
   // Sponsor signs
   const sponsorSignature = aptos.transaction.signAsFeePayer({ signer: sponsor, transaction });
 
   // Submit the transaction to chain
-  const committedTxn = await aptos.transaction.submit.simple({
-    transaction,
-    senderAuthenticator: senderSignature,
+  const committedTxn = await aptos.signAndSubmitTransaction({
+    signer: alice,
     feePayerAuthenticator: sponsorSignature,
+    transaction,
   });
 
   console.log(`Submitted transaction: ${committedTxn.hash}`);