+++ CHANGELOG.md
@@ -4,8 +4,13 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 
 # Unreleased
 
+# 1.29.0 (2024-10-04)
+
 - Remove usage of Buffer.from and replace with TextEncoder for greater compatibility
 - Switch `getAccountCoinAmount` to use a `view` function for more up to date data
+- Add support for federated keyless accounts as defined in AIP-96
+- Add a `signAndSubmitAsFeePayer` function to the API to allow for submission by the fee payer.
+- Add an optional `feePayerAuthenticator` and `feePayer` parameter to `signAndSubmitTransaction` to support signing and submitting in one line.
 
 # 1.28.0 (2024-09-19)
 
@@ -14,9 +19,6 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 - Populate `coinType` for `getAccountCoinAmount` if only `faMetadataAddress` is provided.
 - [`Fix`] `getModuleEventsByEventType` will also account for EventHandle events.
 - [`Hot Fix`] change regex to find object address when using `createObjectAndPublishPackage` move function
-- Add support for federated keyless accounts as defined in AIP-96
-- Add a `signAndSubmitAsFeePayer` function to the API to allow for submission by the fee payer.
-- Add an optional `feePayerAuthenticator` and `feePayer` parameter to `signAndSubmitTransaction` to support signing and submitting in one line.
 
 # 1.27.1 (2024-08-23)